+++ src/lib.rs
@@ -195,9 +195,6 @@ pub use config::{
     configure, Config, DriveStrength, Hysteresis, OpenDrain, PullKeeper, SlewRate, Speed,
 };
 
-#[allow(deprecated)]
-pub use config::{PullKeep, PullKeepSelect, PullUpDown};
-
 /// Re-export of top-level components, without the chip-specific modules.
 ///
 /// `prelude` is to help HAL implementors re-export the `imxrt-iomuxc` APIs
@@ -217,9 +214,6 @@ pub mod prelude {
     pub use crate::config::{
         configure, Config, DriveStrength, Hysteresis, OpenDrain, PullKeeper, SlewRate, Speed,
     };
-
-    #[allow(deprecated)]
-    pub use crate::config::{PullKeep, PullKeepSelect, PullUpDown};
 
     pub use crate::{
         adc, alternate, ccm, clear_sion, consts, flexpwm, gpio, lpi2c, lpspi, lpuart, sai,