+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- **BREAKING** Remove the deprecated pull / keeper configuration API.
+
 ## [0.2.0] 2022-11-30
 
 ### Changed